refactor: move test util
